wip cleaning up && Reorganising the tree structure
