+++ README.md
@@ -1,1 +1,13 @@
 # Project_DA_Java_EN_Come_to_the_Rescue_of_a_Java_Application
+## Sample project for OpenClassroom's fundamentals of Java
+
+1. Lancé un run de la methode __main()__ pour voir ce qui se passe  
+    - `javac src/Project02Eclipse/com/hemebiotech/analytics/AnalyticsCounter.java`
+    - `java src/Project02Eclipse/com/hemebiotech/analytics/AnalyticsCounter.java`
+        - ou Run 'AnalyticsCounter.main()' *"Merci mon IDE 😘" *
+    - Résultat: `Exception in thread "main" java.io.FileNotFoundException: symptoms.txt (No such file or directory)`
+2. Réorganiser l'arborescence la structure des dossiers et fichiers de *Project02Eclipse/*
+3. Nettoyer les doublons dans le fichier symptômes.txt
+
+
+
