+++ result.out
@@ -1,0 +1,3 @@
+headache: 0
+rash: 0
+dilated pupils: 2
