+++ src/com/analytics/ReadSymptomDataFromFile.java
@@ -1,4 +1,4 @@
-package com.hemebiotech.analytics;
+package com.analytics;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
@@ -8,14 +8,12 @@ import java.util.List;
 
 /**
  * Simple brute force implementation
- *
  */
 public class ReadSymptomDataFromFile implements ISymptomReader {
 
-	private String filepath;
+	private final String filepath;
 	
 	/**
-	 * 
 	 * @param filepath a full or partial path to file with symptom strings in it, one per line
 	 */
 	public ReadSymptomDataFromFile (String filepath) {
