+++ src/com/hemebiotech/AnalyticsCounter.java
@@ -1,23 +1,23 @@
-package com.hemebiotech.analytics;
+package com.hemebiotech;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.FileWriter;
 
 public class AnalyticsCounter {
-	private static int headacheCount = 0;	// initialize to 0
-	private static int rashCount = 0;		// initialize to 0
-	private static int pupilCount = 0;		// initialize to 0
+	private static final int headacheCount = 0;
+	private static int rashCount = 0;
+	private static int pupilCount = 0;
 	
 	public static void main(String args[]) throws Exception {
-		// first get input
+
 		BufferedReader reader = new BufferedReader (new FileReader("symptoms.txt"));
 		String line = reader.readLine();
 
-		int i = 0;	// set i to 0
-		int headCount = 0;	// counts headaches
+		int i = 0;
+		int headCount = 0;
 		while (line != null) {
-			i++;	// increment i
+			i++;
 			System.out.println("symptom from file: " + line);
 			if (line.equals("headache")) {
 				headCount++;
@@ -29,15 +29,13 @@ public class AnalyticsCounter {
 			else if (line.contains("pupils")) {
 				pupilCount++;
 			}
-
-			line = reader.readLine();	// get another symptom
+			line = reader.readLine();
 		}
-		
-		// next generate output
+
 		FileWriter writer = new FileWriter ("result.out");
 		writer.write("headache: " + headacheCount + "\n");
 		writer.write("rash: " + rashCount + "\n");
-		writer.write("dialated pupils: " + pupilCount + "\n");
+		writer.write("dilated pupils: " + pupilCount + "\n");
 		writer.close();
 	}
 }
