+++ symptoms.txt
